feat(TDI-39294): Rename auth package + license
